move errors into dedicated package
