+++ backend/src/main/java/it/andmora/expensesmonitor/backend/domain/errors/WalletNotEmptyException.java
@@ -1,0 +1,7 @@
+package it.andmora.expensesmonitor.backend.domain.errors;
+
+public class WalletNotEmptyException extends Throwable {
+    public WalletNotEmptyException() {
+        super("Cannot delete wallet with payments");
+    }
+}
